pyproject.toml
[tool.poetry]
name = "crimes_api"
version = "0.1.0"
description = ""
authors = ["Denis Zvyagintsev <dizvyagintsev@avito.ru>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
google-cloud-bigquery = "^3.3.5"
fastapi = "^0.85.1"
pytest = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
